viet_summ/requirements.txt
transformers==4.19.1
ruamel.yaml==0.17.21
tensorboardX
pyrouge